+++ crates/tuirealm-textarea/examples/editor.rs
@@ -6,9 +6,6 @@ use std::time::Duration;
 #[cfg(feature = "search")]
 use tui_realm_stdlib::components::Input;
 use tui_realm_stdlib::components::Label;
-// textarea
-#[cfg(feature = "clipboard")]
-use tui_realm_textarea::TEXTAREA_CMD_PASTE;
 use tui_realm_textarea::{
     TEXTAREA_CMD_MOVE_WORD_BACK, TEXTAREA_CMD_MOVE_WORD_FORWARD, TEXTAREA_CMD_NEWLINE,
     TEXTAREA_CMD_REDO, TEXTAREA_CMD_UNDO, TextArea,
@@ -31,7 +28,7 @@ use tuirealm::ratatui::layout::{Constraint, Direction as LayoutDirection, Layout
 use tuirealm::state::State;
 #[cfg(feature = "search")]
 use tuirealm::state::StateValue;
-use tuirealm::terminal::CrosstermTerminalAdapter;
+use tuirealm::terminal::{CrosstermTerminalAdapter, TerminalAdapter, TerminalResult};
 
 // -- message
 #[derive(Debug, PartialEq)]
@@ -61,6 +58,16 @@ struct Model {
 }
 
 impl Model {
+    /// Initialize the Terminal modes.
+    fn init_adapter() -> TerminalResult<CrosstermTerminalAdapter> {
+        let mut adapter = CrosstermTerminalAdapter::new()?;
+        adapter.enable_raw_mode()?;
+        adapter.enter_alternate_screen()?;
+        adapter.enable_bracketed_paste()?;
+
+        Ok(adapter)
+    }
+
     fn new() -> Self {
         // Setup app
         let mut app: Application<Id, Msg, NoUserEvent> = Application::init(
@@ -84,7 +91,7 @@ impl Model {
             app,
             quit: false,
             redraw: true,
-            terminal: CrosstermTerminalAdapter::new().expect("Could not initialize terminal"),
+            terminal: Self::init_adapter().expect("Could not initialize terminal"),
         }
     }
 
@@ -372,14 +379,6 @@ impl AppComponent<Msg, NoUserEvent> for Editor {
                 self.perform(Cmd::Custom(TEXTAREA_CMD_SEARCH_FORWARD));
                 Some(Msg::None)
             }
-            #[cfg(feature = "clipboard")]
-            Event::Keyboard(KeyEvent {
-                code: Key::Char('v'),
-                modifiers: KeyModifiers::CONTROL,
-            }) => {
-                self.perform(Cmd::Custom(TEXTAREA_CMD_PASTE));
-                Some(Msg::None)
-            }
             Event::Keyboard(KeyEvent {
                 code: Key::Char('z'),
                 modifiers: KeyModifiers::CONTROL,
@@ -414,6 +413,10 @@ impl AppComponent<Msg, NoUserEvent> for Editor {
                 code: Key::Function(3),
                 ..
             }) => Some(Msg::ChangeFocus(Id::Search)),
+            Event::Paste(text) => {
+                self.component.paste(text);
+                Some(Msg::None)
+            }
             _ => None,
         }
     }