+++ crates/tuirealm-textarea/examples/single_line.rs
@@ -1,8 +1,5 @@
 use std::time::Duration;
 
-// textarea
-#[cfg(feature = "clipboard")]
-use tui_realm_textarea::TEXTAREA_CMD_PASTE;
 use tui_realm_textarea::{
     TEXTAREA_CMD_MOVE_WORD_BACK, TEXTAREA_CMD_MOVE_WORD_FORWARD, TEXTAREA_CMD_NEWLINE,
     TEXTAREA_CMD_REDO, TEXTAREA_CMD_UNDO, TextArea,
@@ -19,7 +16,7 @@ use tuirealm::props::{
 // tui
 use tuirealm::ratatui::layout::{Constraint, Direction as LayoutDirection, Layout};
 use tuirealm::state::State;
-use tuirealm::terminal::CrosstermTerminalAdapter;
+use tuirealm::terminal::{CrosstermTerminalAdapter, TerminalAdapter, TerminalResult};
 
 // -- message
 #[derive(Debug, PartialEq)]
@@ -43,6 +40,16 @@ struct Model {
 }
 
 impl Model {
+    /// Initialize the Terminal modes.
+    fn init_adapter() -> TerminalResult<CrosstermTerminalAdapter> {
+        let mut adapter = CrosstermTerminalAdapter::new()?;
+        adapter.enable_raw_mode()?;
+        adapter.enter_alternate_screen()?;
+        adapter.enable_bracketed_paste()?;
+
+        Ok(adapter)
+    }
+
     fn new() -> Self {
         // Setup app
         let mut app: Application<Id, Msg, NoUserEvent> = Application::init(
@@ -57,7 +64,7 @@ impl Model {
             app,
             quit: false,
             redraw: true,
-            terminal: CrosstermTerminalAdapter::new().expect("Could not initialize terminal"),
+            terminal: Self::init_adapter().expect("Could not initialize terminal"),
         }
     }
 
@@ -278,14 +285,6 @@ impl AppComponent<Msg, NoUserEvent> for Input {
                 self.perform(Cmd::GoTo(Position::Begin));
                 Some(Msg::None)
             }
-            #[cfg(feature = "clipboard")]
-            Event::Keyboard(KeyEvent {
-                code: Key::Char('v'),
-                modifiers: KeyModifiers::CONTROL,
-            }) => {
-                self.perform(Cmd::Custom(TEXTAREA_CMD_PASTE));
-                Some(Msg::None)
-            }
             Event::Keyboard(KeyEvent {
                 code: Key::Char('z'),
                 modifiers: KeyModifiers::CONTROL,
@@ -309,6 +308,10 @@ impl AppComponent<Msg, NoUserEvent> for Input {
                 ..
             }) => {
                 self.perform(Cmd::Type(*ch));
+                Some(Msg::None)
+            }
+            Event::Paste(text) => {
+                self.component.paste(text);
                 Some(Msg::None)
             }
             _ => None,